adding tests for WASM API: `encode_calldata / decode_calldata`
